+++ model/src/main/scala/com/github/ppotseluev/algorate/Bar.scala
@@ -8,7 +8,7 @@ case class Bar(
     closePrice: Price,
     lowPrice: Price,
     highPrice: Price,
-    volume: Long,
+    volume: Double,
     endTime: OffsetDateTime,
     duration: FiniteDuration
 )
