+++ server/src/main/scala/com/github/ppotseluev/algorate/server/Codecs.scala
@@ -51,7 +51,7 @@ object Codecs {
         closePrice: BigDecimal,
         lowPrice: BigDecimal,
         highPrice: BigDecimal,
-        volume: Long,
+        volume: Double,
         endTimeSeconds: Long,
         endTimeNano: Int,
         duration: FiniteDuration
